# python-fastapi/src/openapi_server/models/work_update_result.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, validator  # noqa: F401
from openapi_server.models.work import Work


class WorkUpdateResult(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    WorkUpdateResult - a model defined in OpenAPI

        updated: The updated of this WorkUpdateResult.
        errored: The errored of this WorkUpdateResult.
    """

    updated: List[Work] = []
    errored: List[Work] = []

    class Config:
        orm_mode = True

WorkUpdateResult.update_forward_refs()